Scoring judges need to be optional in the scheduler.
This is a bit annoying, but at the state level competition scoring
judges may not show up until the last minute and asked to be placed in a
room to score. That means that using them in the scheduling algorithm
isn't as cut and dry as I was hoping. (Which is why it is annoying.
Stupid realities and stuff messing up my beautiful code!!)