Move the package structure around a bit.
Conceptually speaking, I think it makes sense for the matching component
to "own" its own models instead of sharing models with the larger
application. It turned out to be a pretty complex piece of code that
will probably need to have good boundaries in place so it can evolve
independent of whatever we decide to do with the rest of the
application.

Therefore, we're going to be instituting a new project structure that
permits this flexibility.

It's, loosely speaking, what I mentioned above. Different modules own
their models. Other modules may read those models as-is or convert them
to a format more amicable to whatever they want to do, but the models
should ideally only change for the needs of the component they belong
to. So, you can imagine the persistence layer having its own
representation of things that is sensible for persisting and the web UI
probably largely reading the matching models mostly 1:1 because the
information isn't all that different.

Still a bit of a work in progress so I give it about a day before I
break these rules somehow.