Implement DOM event triggering
This code permits us to trigger and handle arbitrary DOM events between the server and client. This
is a useful mechanism for the server to let the client know things have happened without having any
knowledge of what the client actually might want to do with that information. As far as
browser-implemented messages busses go, the DOM event system is actually prettny handy. We can even
register multiple handlers at will for a particular event - and the server need not know anything
about them.